Migrate Extent Report Ver 5
