+++ src/main/java/common/Constant.java
@@ -2,8 +2,4 @@ package common;
 
 public class Constant {
 
-    public static final String RAILWAY_URL = "http://railway2.somee.com";
-    public static final String USERNAME = "hoang@gmail.com";
-    public static final String PASSWORD = "         ";
-
 }
