+++ src/main/java/testcase/BaseTest.java
@@ -2,11 +2,10 @@ package testcase;
 
 import static common.Driver.webDriver;
 
-import com.relevantcodes.extentreports.ExtentReports;
-import com.relevantcodes.extentreports.ExtentTest;
 import common.Constant;
 import common.Driver;
-import common.Utilities;
+import common.Utis;
+import data.Account;
 import org.testng.ITestContext;
 import org.testng.annotations.*;
 import org.testng.annotations.Test;
@@ -20,40 +19,39 @@ import java.lang.reflect.Method;
 public class BaseTest {
 
     protected HomePage homePage = new HomePage();
-    private static final String reportPath = Utilities.generatReportPath();
+    private static final String reportPath = Utis.generateReportPath();
+    private final String URL = System.getenv("URL");
+    protected final Account account = new Account(System.getenv("USERNAME"), System.getenv("PASSWORD"));
 
-    protected static ExtentReports reports;
-    protected static ExtentTest test;
+//    protected static ExtentReports reports;
+//    protected static ExtentTest test;
 
-    @BeforeClass
+    @BeforeSuite
     public static void beforeClass() {
-        reports = new ExtentReports(reportPath);
+//        reports = new ExtentReports(reportPath);
     }
 
     @BeforeMethod
-    public void beforeMethod(ITestContext context, Method method) {
+    @Parameters(value = {"browser", "path"})
+    public void beforeMethod(String browser, String path, Method method) {
         Test t = method.getAnnotation(Test.class);
-        test = reports.startTest(String.format("%s - %s", method.getName(), t.description()));
+//        test = reports.startTest(String.format("%s - %s", method.getName(), t.description()));
 
-        String browser = context.getCurrentXmlTest().getParameter("browser");
-        String pathToDriver = context.getCurrentXmlTest().getParameter("path");
-        Driver.initDriver(browser, pathToDriver);
-//        System.setProperty("webdriver.gecko.driver", System.getProperty("user.dir") + "/driver/geckodriver.exe");
-//        webDriver = new FirefoxDriver();
-        webDriver.get(Constant.RAILWAY_URL);
+        Driver.initDriver(browser, path);
+        webDriver.get(URL);
         webDriver.manage().window().maximize();
     }
 
     @AfterMethod
     public void afterMethod() {
-        reports.endTest(test);
+//        reports.endTest(test);
         webDriver.quit();
     }
 
 
-    @AfterClass
+    @AfterSuite
     public static void afterClass() throws IOException {
-        reports.flush();
+//        reports.flush();
 
         // Open the report after done
         File htmlFile = new File(reportPath);
