+++ xml/testng.xml
@@ -1,11 +1,26 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd">
-<suite name="Railway Testcase">
+<suite name="Railway Testcase" parallel="false">
+    <parameter name="path" value="\driver\chromedriver.exe"/>
+    <parameter name="browser" value="chrome"/>
     <test name="Set 1">
-        <parameter name="path" value="\driver\chromedriver.exe"/>
-        <parameter name="browser" value="chrome"/>
         <classes>
-            <class name="testcase.Set1"/>
+            <class name="testcase.TC01"></class>
+            <class name="testcase.TC02"></class>
+            <class name="testcase.TC03"></class>
+            <class name="testcase.TC04"></class>
+            <class name="testcase.TC05"></class>
+            <class name="testcase.TC06"></class>
+            <class name="testcase.TC07"></class>
+            <class name="testcase.TC08"></class>
+            <class name="testcase.TC09"></class>
+            <class name="testcase.TC10"></class>
+            <class name="testcase.TC11"></class>
+            <class name="testcase.TC12"></class>
+            <class name="testcase.TC13"></class>
+            <class name="testcase.TC14"></class>
+            <class name="testcase.TC15"></class>
+            <class name="testcase.TC16"></class>
         </classes>
     </test>
 </suite>
