+++ xml/testng_edge.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd">
+<suite name="Practice Suite">
+    <test name="Test Basics 1">
+        <parameter name="path" value="\driver\msedgedriver.exe"/>
+        <parameter name="browser" value="edge"/>
+        <classes>
+            <class name="testcase.Set1"/>
+        </classes>
+    </test>
+</suite>
